Move ui block shadows into the shader

+++ src/shaders/ui_overlay_fragment.glsl
@@ -5,14 +5,14 @@ precision lowp sampler2DArray;
 
 uniform sampler2DArray u_Texture;
 
-flat in float v_is_block;
+flat in uint v_is_block;
 in vec3 v_TexCoordBlock;
 in vec4 v_Tint;
 
 layout( location = 0 ) out vec4 color;
 
 void main( ) {
-    if ( v_is_block != 0.0 ) {
+    if ( v_is_block != 0u ) {
         color = texture( u_Texture, v_TexCoordBlock );
         color *= v_Tint;
     } else {
